wip: latest llama with Metal
not fully baked yet:

* Currently need to ask for a Metal build with cmake -DLLAMA_METAL=ON
* Metal *only* works with q4_0, any other files will fail to load on a
  Metal build - need to somehow, ugh, detect beforehand if the file is
  anything other than q4_0 to pass n_gpu_layers=0 to disable Metal
* `ggml-metal.metal` needs to be included in `Contents/Resources` in the
  app bundle and the documented way to do that in CMake isn't doing the
  trick, but it works after manually copying it in

---

also this of only curently does anything for llama ggmlv3 models. Metal
(and ggml GPU offload generally) requires some additional support on the
model code side to move stuff in and out of GPU memory, and in addition
to the current q4_0 quant limitation types, only a subset (the subset
needed to run llama) of the ggml ops are currently implemented in Metal.
Currently missing for supporting additional models are

 * GELU activation (most models use this, but llama uses swish)
 * ALiBi positional bias (MPT, Replit)
 * regular old layernorm (llama uses RMSNorm)